requirements.txt
txtai[all]
python-dotenv
PyPDF2
langchain[all]
streamlit
sentence-transformers
torch
transformers
sentencepiece
accelerate